Revert "Translation widget client (#32)"
This reverts commit 754f728ae009aa69a0ded1882d7028ead6065586.
