+++ client/extensions/auto-translator/src/components/select.tsx
@@ -18,6 +18,7 @@ type SelectProps<T> = Omit<
   helpers?: FieldHelperProps<T>;
   value?: string;
   onChange?: (newValue: string) => void;
+  [key: string]: any;
 };
 
 export const Select = <T,>({
