+++ client/extensions/auto-translator/src/components/text-input.tsx
@@ -10,11 +10,10 @@ import { RecursiveKeyOf } from "../formik-utilties";
 
 type TextInputProps<T> = React.InputHTMLAttributes<HTMLInputElement> & {
   label: string;
-  field?: FieldInputProps<T>;
-  meta?: FieldMetaProps<T>;
-  helpers?: FieldHelperProps<T>;
-  readonly?: boolean;
-  onChange?: (newValue: string) => void;
+  field: FieldInputProps<T>;
+  meta: FieldMetaProps<T>;
+  helpers: FieldHelperProps<T>;
+  [key: string]: any;
 };
 
 export const TextInput = <T,>({
@@ -22,7 +21,6 @@ export const TextInput = <T,>({
   field,
   meta,
   helpers,
-  onChange,
   ...props
 }: TextInputProps<T>) => {
   return (
@@ -31,13 +29,12 @@ export const TextInput = <T,>({
       {...props}
       type="text"
       label={label}
-      boxedLable
-      boxedStyle
+      boxedLable={true}
+      boxedStyle={true}
       size="medium"
       value={field?.value as string}
       onChange={(newValue) => {
-        if (onChange) onChange(newValue);
-        else if (helpers) helpers.setValue(newValue as T);
+        helpers.setValue(newValue as T);
       }}
       error={meta?.error ? meta.error : undefined}
     />
@@ -53,7 +50,7 @@ export const FormTextInput = <T,>({
   label,
   ...props
 }: FormTextInputProps<T>) => {
-  const [field, meta, helpers] = useField<T>({ name });
+  const [field, meta, helpers] = useField<T>(name);
 
   return (
     <TextInput<T>
