+++ client/extensions/auto-translator/translations/fr_CA.po
@@ -5,8 +5,20 @@ msgstr ""
 msgid "Translate"
 msgstr "Traduire"
 
+msgid "Auto"
+msgstr "Automatique"
+
 msgid "Cancel"
 msgstr "Annuler"
+
+msgid "Apply"
+msgstr "Appliquer"
+
+msgid "Translation"
+msgstr "Traduction"
+
+msgid "Version"
+msgstr "Version"
 
 msgid "Original"
 msgstr "Originale"
@@ -26,8 +38,32 @@ msgstr "Titre Étendu"
 msgid "body HTML"
 msgstr "corps HTML"
 
+msgid "Photos"
+msgstr "Des Photos"
+
+msgid "Caption"
+msgstr "Légende"
+
 msgid "Writethru"
 msgstr "Lead"
+
+msgid "Writethru 1"
+msgstr "Lead 1"
+
+msgid "Writethru 2"
+msgstr "Lead 2"
+
+msgid "Current"
+msgstr "Actuel"
+
+msgid "Type"
+msgstr "Type"
+
+msgid "From"
+msgstr "De"
+
+msgid "To"
+msgstr "En"
 
 msgid "English"
 msgstr "Anglais"
@@ -38,8 +74,8 @@ msgstr "Français"
 msgid "Compare"
 msgstr "Comparer"
 
-msgid "Difference (Diff)"
-msgstr "Différence (Diff)"
+msgid "Diff"
+msgstr "Diff"
 
 msgid "Inserted"
 msgstr "Inséré"
@@ -53,26 +89,8 @@ msgstr "Traduire Automatique"
 msgid "Apply Translation"
 msgstr "Appliquer Traduction"
 
-msgid "Confirm Translate"
-msgstr "Confirmer Traduire"
-
-msgid "Yes"
-msgstr "Oui"
-
-msgid "No"
-msgstr "Non"
-
-msgid "By clicking \"Translate\" again, changes to the current manual translation will be lost. Are you sure you wish to proceed?"
-msgstr "En cliquant à nouveau sur « Traduire », les modifications apportées à la traduction actuelle du manuel seront perdues. Êtes-vous sûr de vouloir continuer ?"
-
-msgid "Auto Translate"
-msgstr "Traduire Automatique"
-
-msgid "Apply Translation"
-msgstr "Appliquer Traduire"
-
-msgid "Translation Type (Engine)"
-msgstr "Type de Traduction (Moteur)"
+msgid "Translation Type"
+msgstr "Type de Traduction"
 
 msgid "Translate From"
 msgstr "Traduire de"
@@ -80,38 +98,5 @@ msgstr "Traduire de"
 msgid "Translate To"
 msgstr "Traduire en"
 
-msgid "Compare Writethru Difference"
-msgstr "Comparer Lead Différence"
-
-msgid "Search"
-msgstr "Recherche"
-
-msgid "Replace"
-msgstr "Remplacer"
-
-msgid "Replace All"
-msgstr "Remplacer Tout"
-
-msgid "Translation Widget"
-msgstr "Widget de Traduction"
-
-msgid "Current Story"
-msgstr "Histoire Actuelle"
-
-msgid "Version (Original Content)"
-msgstr "Version (Contenu Original)"
-
-msgid "Version (Translated Content)"
-msgstr "Version (Contenu Traduit)"
-
-msgid "Headline may have a maximum character length of"
-msgstr "Le Titre peut avoir une longueur maximale de caractères de"
-
-msgid "Extended Headline may have a maximum character length of"
-msgstr "Le Titre Étendu peut avoir une longueur maximale de caractères de"
-
-msgid "Clear"
-msgstr "Effacer"
-
-msgid "Correction #"
-msgstr "Corrigè #"
+msgid "Compare Writethru Diff"
+msgstr "Comparer Lead Diff"
