+++ server/settings.py
@@ -176,11 +176,6 @@ SCHEMA = {
         "usageterms": {"required": False},
         "ednote": {"required": False},
     },
-    "Story": {
-        "headline": {"required": False},
-        "headline_extended": {"required": False},
-        "body_html": {"required": False},
-    },
 }
 
 # editor for images, video, audio
@@ -206,7 +201,6 @@ EDITOR = {
 }
 
 SCHEMA["audio"] = SCHEMA["video"]
-SCHEMA["Story"] = SCHEMA["Story"]
 EDITOR["audio"] = EDITOR["video"]
 
 # if there is picture/audio/video content type defined in data/content_types
